chore : Update aux drawers
