+++ src/constants/idHtmlReference.constants.ts
@@ -1,2 +1,3 @@
 export const HEADER_ID = '#header';
 export const HEADER_FILTER_ID = '#header-filters';
+export const AUX_DRAWER_ID = '#aux-sidebar';
