+++ src/modules/vulnerability/components/drawer/ReportDetailDrawer.vue
@@ -1,0 +1,56 @@
+<template>
+  <div style="background-color: #313541; height: 100%" class="q-pa-md">
+    <p class="text-white text-h6">SEVERITY</p>
+    <div class="row q-mb-sm">
+      <div class="col-10">
+        <span class="card-red">CRITICAL</span>
+      </div>
+      <div class="col-2 text-white text-weight-bold">
+        {{ vulnerabilities.severity_counts.critical }}
+      </div>
+    </div>
+    <div class="row q-mb-sm">
+      <div class="col-10">
+        <span class="card-orange">HIGH</span>
+      </div>
+      <div class="col-2 text-white text-weight-bold">
+        {{ vulnerabilities.severity_counts.high }}
+      </div>
+    </div>
+    <div class="row q-mb-sm">
+      <div class="col-10"><span class="card-yellow">MEDIUM</span></div>
+      <div class="col-2 text-white text-weight-bold">
+        {{ vulnerabilities.severity_counts.medium }}
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-10"><span class="card-green">LOW</span></div>
+      <div class="col-2 text-white text-weight-bold">
+        {{ vulnerabilities.severity_counts.low }}
+      </div>
+    </div>
+    <q-separator class="q-my-sm text-white bg-white" />
+    <div class="row text-white text-weight-bold">
+      <div class="col-10">TOTAL</div>
+      <div class="col-2">{{ totalSeverity }}</div>
+    </div>
+  </div>
+</template>
+
+<script lang="ts" setup>
+  import { ScanVulnerabilitesResponse } from 'src/models';
+  import { computed, PropType } from 'vue';
+
+  const props = defineProps({
+    vulnerabilities: {
+      type: Object as PropType<ScanVulnerabilitesResponse>,
+      required: true
+    }
+  });
+
+  const totalSeverity = computed(() => {
+    return Object.values(props.vulnerabilities.severity_counts).reduce(
+      (aux = 0, item) => (aux = aux + item)
+    );
+  });
+</script>
