+++ src/modules/vulnerability/components/drawer/ReportDetailVulnerabilityDrawer.vue
@@ -1,0 +1,111 @@
+<template>
+  <div style="background-color: #313541; height: 100%; overflow-y: auto" class="q-pa-md">
+    <p class="text-white text-h6">DETAILS</p>
+    <q-list class="rounded-borders">
+      <q-expansion-item
+        expand-separator
+        label="DATE"
+        dense
+        header-style="padding:0;font-weight:600"
+        class="text-white vul-container"
+      >
+        <q-item class="vul-info">
+          <span class="vul-title">Published : </span>
+          <span class="vul-des"> {{ vulnerability.date?.published }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Last Updated :</span>
+          <span class="vul-des"> {{ vulnerability.date?.last_updated }}</span>
+        </q-item>
+      </q-expansion-item>
+      <q-expansion-item
+        expand-separator
+        label="PLUGIN"
+        dense
+        class="text-white"
+        header-style="padding:0;font-weight:600"
+      >
+        <q-item class="vul-info">
+          <span class="vul-title">CPE :</span>
+          <span class="vul-des"> {{ vulnerability.plugin?.cpe }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Severity :</span>
+          <span class="vul-des"> {{ vulnerability.plugin?.severity }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Version :</span>
+          <span class="vul-des"> {{ vulnerability.plugin?.version }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Family :</span>
+          <span class="vul-des"> {{ vulnerability.plugin?.family }}</span>
+        </q-item>
+      </q-expansion-item>
+      <q-expansion-item
+        expand-separator
+        label="VPR KEY D."
+        dense
+        class="text-white"
+        header-style="padding:0;font-weight:600"
+      >
+        <q-item class="vul-info">
+          <span class="vul-title">Threat Intensity :</span>
+          <span class="vul-des"> {{ vulnerability.vpr_key_d?.threat_intensity }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Exploit Code Maturity :</span>
+          <span class="vul-des"> {{ vulnerability.vpr_key_d?.exploit_code_maturity }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Age of Vul :</span>
+          <span class="vul-des"> {{ vulnerability.vpr_key_d?.age_of_vuln }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Product Coverage :</span>
+          <span class="vul-des"> {{ vulnerability.vpr_key_d?.product_coverage }}</span>
+        </q-item>
+      </q-expansion-item>
+      <q-expansion-item
+        expand-separator
+        label="RISK INFO"
+        dense
+        class="text-white"
+        header-style="padding:0;font-weight:600"
+      >
+        <q-item class="vul-info">
+          <span class="vul-title">Risk Score:</span>
+          <span class="vul-des"> {{ vulnerability.risk?.risk_score }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Availability Impact :</span>
+          <span class="vul-des"> {{ vulnerability.risk?.availability_impact }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">Integrity Impact :</span>
+          <span class="vul-des"> {{ vulnerability.risk?.integrity_impact }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">CVSS v3 Base :</span>
+          <span class="vul-des"> {{ vulnerability.risk?.cvss_v3_base }}</span>
+        </q-item>
+        <q-item class="vul-info">
+          <span class="vul-title">CVSS v3 Vector :</span>
+          <span class="vul-des"> {{ vulnerability.risk?.cvss_v3_vector }}</span>
+        </q-item>
+      </q-expansion-item>
+    </q-list>
+  </div>
+</template>
+
+<script lang="ts" setup>
+  import { ScanVulnerability } from 'src/models';
+  import { PropType } from 'vue';
+
+  defineProps({
+    vulnerability: {
+      required: true,
+      type: Object as PropType<ScanVulnerability>
+    }
+  });
+</script>
