+++ src/modules/vulnerability/pages/ReportDetailPage.vue
@@ -3,43 +3,8 @@
     <Teleport :to="HEADER_ID">
       <report-detail-header :vulnerability-alias="vulnerabilities.alias" />
     </Teleport>
-    <Teleport to="#aux-sidebar">
-      <div style="background-color: #313541; height: 100%" class="q-pa-md">
-        <p class="text-white text-h6">SEVERITY</p>
-        <div class="row q-mb-sm">
-          <div class="col-10">
-            <span class="card-red">CRITICAL</span>
-          </div>
-          <div class="col-2 text-white text-weight-bold">
-            {{ vulnerabilities.severity_counts.critical }}
-          </div>
-        </div>
-        <div class="row q-mb-sm">
-          <div class="col-10">
-            <span class="card-orange">HIGH</span>
-          </div>
-          <div class="col-2 text-white text-weight-bold">
-            {{ vulnerabilities.severity_counts.high }}
-          </div>
-        </div>
-        <div class="row q-mb-sm">
-          <div class="col-10"><span class="card-yellow">MEDIUM</span></div>
-          <div class="col-2 text-white text-weight-bold">
-            {{ vulnerabilities.severity_counts.medium }}
-          </div>
-        </div>
-        <div class="row">
-          <div class="col-10"><span class="card-green">LOW</span></div>
-          <div class="col-2 text-white text-weight-bold">
-            {{ vulnerabilities.severity_counts.low }}
-          </div>
-        </div>
-        <q-separator class="q-my-sm text-white bg-white" />
-        <div class="row text-white text-weight-bold">
-          <div class="col-10">TOTAL</div>
-          <div class="col-2">{{ totalSeverity }}</div>
-        </div>
-      </div>
+    <Teleport :to="AUX_DRAWER_ID">
+      <report-detail-drawer :vulnerabilities="vulnerabilities" />
     </Teleport>
   </template>
 
@@ -61,13 +26,14 @@
   import type { ScanVulnerabilitesResponse, ScanVulnerability } from 'src/models';
   import { ReportService } from 'src/services';
   import type { Ref } from 'vue';
-  import { computed, onMounted, onUnmounted, ref } from 'vue';
+  import { onMounted, onUnmounted, ref } from 'vue';
   import { useRoute } from 'vue-router';
   import { useQuasar } from 'quasar';
   import VulnerabilityCard from 'src/components/report/VulnerabilityCard.vue';
   import { useMiscStore } from 'vulnerability/stores/misc';
   import ReportDetailHeader from 'vulnerability/components/header/ReportDetailHeader.vue';
-  import { HEADER_ID } from 'src/constants/idHtmlReference.constants';
+  import ReportDetailDrawer from 'vulnerability/components/drawer/ReportDetailDrawer.vue';
+  import { HEADER_ID, AUX_DRAWER_ID } from 'src/constants/idHtmlReference.constants';
 
   const route = useRoute();
   const scanId = route.params.id as string;
@@ -76,12 +42,6 @@
   const isMounted = ref(false);
   const $q = useQuasar();
   const miscStore = useMiscStore();
-
-  const totalSeverity = computed(() => {
-    return Object.values(vulnerabilities.value.severity_counts).reduce(
-      (aux = 0, item) => (aux = aux + item)
-    );
-  });
 
   function handleCommentUpdated(payload: { vulID: number; newComment: string }) {
     const vulnerabilityToUpdate = vulnerabilities.value.vulnerabilities.find(
